+++ packages/devtools_extensions/example/foo/packages/foo_devtools_extension/lib/src/devtools_extension_api_example.dart
@@ -27,7 +27,7 @@ class CallingDevToolsExtensionsAPIsExample extends StatelessWidget {
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         Text(
-          '3. Example of calling DevTools extension APIs',
+          '4. Example of calling DevTools extension APIs',
           style: Theme.of(context).textTheme.titleMedium,
         ),
         const PaddedDivider.thin(),